+++ apps/wsb-task/src/commands/cargo/build_apps/runner.rs
@@ -1,11 +1,11 @@
-use crate::commands::cargo_build;
+use crate::commands::build_apps;
 use crate::commands::support::{mac, should, CanInsertCC, Definable};
 use crate::core::targets::{BuildTarget, LinuxArmV7, LinuxX86, MacX86};
 use crate::TaskResult;
 use shellwork::core::command;
 use shellwork::core::command::{Runner, Unprepared};
 
-fn base_runner<T>(params: &cargo_build::Params<T>) -> Runner<Unprepared>
+fn base_runner<T>(params: &build_apps::Params<T>) -> Runner<Unprepared>
 where
     T: BuildTarget,
 {
@@ -20,32 +20,32 @@ where
 mod linux_x86 {
     use super::*;
 
-    impl Definable for cargo_build::Params<LinuxX86> {
+    impl Definable for build_apps::Params<LinuxX86> {
         fn define(&self) -> TaskResult<Runner<Unprepared>> {
             Ok(base_runner(self))
         }
     }
-    impl should::Run for cargo_build::Params<LinuxX86> {}
+    impl should::Run for build_apps::Params<LinuxX86> {}
 }
 
 mod linux_arm_v7 {
     use super::*;
 
-    impl Definable for cargo_build::Params<LinuxArmV7> {
+    impl Definable for build_apps::Params<LinuxArmV7> {
         fn define(&self) -> TaskResult<Runner<Unprepared>> {
             self.with_cc(base_runner)
         }
     }
-    impl should::Run for cargo_build::Params<LinuxArmV7> {}
+    impl should::Run for build_apps::Params<LinuxArmV7> {}
 }
 
 mod mac_x86 {
     use super::*;
 
-    impl Definable for cargo_build::Params<MacX86> {
+    impl Definable for build_apps::Params<MacX86> {
         fn define(&self) -> TaskResult<Runner<Unprepared>> {
             self.with_cc(base_runner)
         }
     }
-    impl mac::RunMaybe for cargo_build::Params<MacX86> {}
+    impl mac::RunMaybe for build_apps::Params<MacX86> {}
 }
