+++ apps/wsb-task/src/commands/cargo/mod.rs
@@ -1,2 +1,2 @@
+pub mod build_apps;
 pub mod build_pilot;
-pub mod cargo_build;
